Finished the main program.
The main program runs but needs to be tested. The rounding feature might not work perfectly since some
rounded value end up causing the aspect ratio to be thrown off more than rounding it a different way.

For example: a 15 x 20 dimension at mdpi with the standard rounding function produces a result of 22 x 30
at hdpi. A better result would be 23 x 31 or 22 x 29 since they have  a closer aspect ratio to the original.
This might not be a big enough difference to matter but needs to be looked into more.